+++ config/routes.rb
@@ -104,8 +104,9 @@ Rails.application.routes.draw do
     namespace :v2 do
       resources :apidocs, only: %i(index)
       resources :plips, only: %i(index)
-      resources :petitions, only: [:info] do
+      resources :petitions, only: [:info, :signers] do
         get :info
+        get :signers
       end
     end
   end