+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* [PR #145]: Adding signers widget on petition's page (Issue 109)
 * [PR #143]: Adding oath authentication for partners API (Issue 127)
  - Run `rake db:migrate`
 * [PR #142]: Removing user field requirements (Issue 126)